EmotionPix/requirements.txt
Flask
Flask-Session
python-dotenv
requests
opencv-python-headless
numpy
email-validator
Werkzeug
supabase
fer
